Working directio on Linux for RrdCachedFileBackend

+++ junit/jrds/caching/TestFilePage.java
@@ -1,6 +1,7 @@
 package jrds.caching;
 
 import java.io.File;
+import java.io.FileDescriptor;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -11,12 +12,16 @@ import jrds.Tools;
 
 import org.apache.log4j.Level;
 import org.apache.log4j.Logger;
+import org.junit.After;
 import org.junit.Assert;
+import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
 public class TestFilePage {
     static final private Logger logger = Logger.getLogger(TestRrdCachedFileBackend.class);
+
+    static final private File testFile = new File("tmp/testfilepage");
 
     @BeforeClass
     static public void configure() throws IOException, ParserConfigurationException {
@@ -25,12 +30,20 @@ public class TestFilePage {
         RrdCachedFileBackendFactory.loadDirect(new File("build/native"));
     }
 
+    @Before
+    public void initialize() {
+        if(testFile.exists())
+            testFile.delete();        
+    }
+    
+    @After
+    public void finish() {
+        if(testFile.exists() && ! logger.isTraceEnabled())
+            testFile.delete();        
+    }
+
     @Test
     public void testWrite() throws IOException {
-        File testFile = new File("tmp/testfilepage");
-        if(testFile.exists())
-            testFile.delete();
-        
         FilePage page = new FilePage(0);
         page.load(testFile, 0);
         page.write(0, getClass().getName().getBytes());
@@ -40,15 +53,10 @@ public class TestFilePage {
         byte[] b = new byte[(int) testFile.length()];
         in.read(b);
         Assert.assertEquals("read does not match write", getClass().getName().trim(), new String(b).trim());
-        testFile.delete();
     }
 
     @Test
     public void testRead() throws IOException {
-        File testFile = new File("tmp/testfilepage");
-        if(testFile.exists())
-            testFile.delete();
-        
         FileOutputStream in = new FileOutputStream(testFile);
         in.write(getClass().getName().getBytes());
         in.flush();
@@ -60,6 +68,30 @@ public class TestFilePage {
         page.read(0, b);
 
         Assert.assertEquals("write does not match read", getClass().getName().trim(), new String(b).trim());
+    }
+    
+    @Test(expected=IOException.class)
+    public void testFailed() throws IOException {
+        try {
+            FilePage.prepare_fd(testFile.getCanonicalPath(), new FileDescriptor(), true);
+        } catch (IOException e) {
+            logger.debug("Caught exception: " + e.getMessage());
+            throw e;
+        }
+    }
+
+    @Test(expected=IOException.class)
+    public void testFailed2() throws IOException {
+        FileOutputStream out = new FileOutputStream(testFile);
+        out.write(0);
+        out.close();
+        testFile.setReadOnly();
+        try {
+            FilePage.prepare_fd(testFile.getCanonicalPath(), new FileDescriptor(), false);
+        } catch (IOException e) {
+            logger.debug("Caught exception: " + e.getMessage());
+            throw e;
+        }
     }
 
 }
